envirozen/controller/requirements.txt
requests==2.26.0
RPi.GPIO==0.7.1